+++ packages/ListBoxWithTags/stories/ListBoxWithTags.sizes.screener.stories.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { getStoryName } from "storybook/storyTree";
 import Sizes from "./examples/Sizes";
+import Footer from "./examples/UncontrolledFooter";
 
 const storyName = getStoryName("ListBoxWithTags");
 
@@ -9,3 +10,5 @@ export default {
 };
 
 export const ListBoxWithTagsScreener = () => <Sizes />;
+
+export const ListBoxWithTagsFooterScreener = () => <Footer isOpen />;