+++ packages/MockEndpoints/src/MockEndpoints.js
@@ -1,11 +1,24 @@
+import React from "react";
 import fetchMock from "fetch-mock";
 
 export default fetchMock;
 
-export function mockEndpoints(handlers) {
-  // validate that the handlers have a certain shape (url, responseData)
-  handlers.forEach(handler => {
-    fetchMock.get(handler.url, handler.responseData);
+fetchMock.config.overwriteRoutes = true;
+
+function validateEndpoint(endpoint) {
+  if (!("url" in endpoint)) {
+    console.error('MockEndpoints.js: Each endpoint object must have a "url" key.');
+  }
+
+  if (!("response" in endpoint)) {
+    console.error('MockEndpoints.js: Each endpoint object must have a "response" key.');
+  }
+}
+
+function mockEndpoints(endpoints) {
+  endpoints.forEach(endpoint => {
+    validateEndpoint(endpoint);
+    fetchMock.get(endpoint.url, endpoint.response);
   });
 
   fetchMock.catch(url => {
@@ -13,4 +26,17 @@ export function mockEndpoints(handlers) {
     console.error(error);
     return { error };
   });
+}
+
+export function useMockEndpoints(endpoints) {
+  const [endpointsAreMocked, setEndpointsAreMocked] = React.useState(false);
+
+  React.useEffect(() => {
+    mockEndpoints(endpoints);
+    setEndpointsAreMocked(true);
+  }, [endpoints]);
+
+  mockEndpoints(endpoints);
+
+  return { endpointsAreMocked };
 }