+++ packages/Tabs/src/types.js
@@ -1,4 +1,12 @@
 import * as constants from "@paprika/constants/lib/Constants";
 
-export const PRIMARY = constants.kind.PRIMARY;
-export const SECONDARY = constants.kind.SECONDARY;
+export default {
+  kind: {
+    PRIMARY: constants.kind.PRIMARY,
+    SECONDARY: constants.kind.SECONDARY,
+  },
+  size: {
+    MEDIUM: constants.size.MEDIUM,
+    LARGE: constants.size.LARGE,
+  },
+};