+++ packages/Breadcrumbs/CHANGELOG.md
@@ -23,3 +23,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed the prop type on `<Breadcrumbs.Link>`.
+
+## [1.1.6-alpha.0] - 2021-01-21
+
+### Fixed
+
+- Fixed the potential mapping children error.