+++ packages/ListBoxBrowser/CHANGELOG.md
@@ -11,3 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Moved from ListboxBrowser
+
+## [1.0.9-alpha.0] - 2021-01-21
+
+### Fixed
+
+- Fixed the potential mapping children error.