+++ packages/NotificationCard/CHANGELOG.md
@@ -9,3 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.0.2] - 2020-09-25
 
 - Added `<NotificationCard />` component.
+
+## [1.0.6-alpha.0] - 2021-01-21
+
+### Fixed
+
+- Fixed the potential mapping children error.